requirements.txt
colormath==3.0.0
fitz==0.0.1.dev2
numpy==1.21.1
opencv_contrib_python==4.4.0.46
opencv_python==4.2.0.32
opencv_python_headless==4.4.0.40
pandas==1.3.5
Pillow==9.0.1
